backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2

langchain==0.2.14
langchain-community==0.2.12
pypdf==4.3.1

sentence-transformers==3.0.1

chromadb==0.5.5

groq==0.15.0        
python-dotenv==1.0.1